% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_conf.R
\name{boot_conf}
\alias{boot_conf}
\title{Calculates the 95\% confidence interval using boostrapping}
\usage{
boot_conf(df, bootstrap_iterations, f, ...)
}
\arguments{
\item{df}{a data.frame containing the clumnes phs, age, and status}

\item{bootstrap_iterations}{number of bootstrap iterations to run.}

\item{f}{the name of a function to perform bootstrapping on.}

\item{...}{arguments to be passed on to \code{f()}.}
}
\value{
A list of confidence intervals
}
\description{
Internal function. Not intended for users. This function can return the 95\%
confidence intervals for any statistic so long as the function passed to \code{f}
returns a single value of that statistic.
}
