% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_confint.R
\name{boot_confint}
\alias{boot_confint}
\title{Calculates the 95\% confidence interval using boostrapping}
\usage{
boot_confint(df, B, f, ...)
}
\arguments{
\item{df}{a data.frame containing the clumnes phs, age, and status}

\item{B}{number of bootstrap iterations to run.}

\item{f}{the name of a function to perform bootstrapping on.}

\item{...}{arguments to be passed on to \code{f()}.}
}
\value{
A list of confidence intervals
}
\description{
Internal function. Not intended for users.
}
\examples{

quantiles <- boot_confint(df, B, calc_or, or_age, upper_quantile, lower_quantile)

}
