% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cindex.R
\name{calc_cindex}
\alias{calc_cindex}
\title{Returns the concordance index for a dataset using a coxph fit}
\usage{
calc_cindex(df, swc = FALSE, swc_popnumcases = NULL, swc_popnumcontrols = NULL)
}
\arguments{
\item{df}{a data.frame containing the columns phs, age, and status}

\item{swc}{logical. if \code{TRUE} performs sample weight correction}

\item{swc_popnumcases}{an optional integer specifying the number of cases in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{an optional integer specifying the number of controls in a reference population for sample weight correction. Required if swc = \code{TRUE}.}
}
\value{
A numeric hazard ratio
}
\description{
Internal function. Not intended for users.
}
