% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hr.R
\name{calc_hr}
\alias{calc_hr}
\title{Calculates the hazard ratio for a dataset using a coxph fit}
\usage{
calc_hr(
  df,
  lower_interval,
  upper_interval,
  swc = FALSE,
  swc_popnumcases = NULL,
  swc_popnumcontrols = NULL
)
}
\arguments{
\item{df}{a data.frame containing the clumnes phs, age, and status}

\item{lower_interval}{a vector specifying the quantiles of the lower interval. If a single value is given, that will be used as the upper quantile and the lower quantile will be \code{-Inf}. If a vector of length 2 is provided then these will be used as the lower and upper quantiles of the interval (e.g., \code{c(0.30, 0.70)}). The default is \code{0.2}.}

\item{upper_interval}{a vector specifying the quantiles of the upper interval. If a single value is given, that will be used as the lower quantile and the upper quantile will be \code{Inf}. If a vector of length 2 is provided then these will be used as the lower and upper quantiles of the interval (e.g., \code{c(0.80, 0.98)}). The default is \code{0.80}.}

\item{swc}{logical. if \code{TRUE} performs sample weight correction}

\item{swc_popnumcases}{an optional integer specifying the number of cases in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{an optional integer specifying the number of controls in a reference population for sample weight correction. Required if swc = \code{TRUE}.}
}
\value{
A numeric hazard ratio
}
\description{
Internal function. Not intended for users.
}
