% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookup.R
\name{create_lookup_table}
\alias{create_lookup_table}
\title{Creates a lookup table for individual prediction of risk based on PHS}
\usage{
create_lookup_table(data = NULL, phs = "phs", age = "age", status = "status")
}
\arguments{
\item{data}{an optional data.frame containing the variables for phs, age, and status}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject or the unquoted name of a vector containing these values. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject or the unquoted name of a vector containing these values. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event) or the unquoted name of a vector containing these values. The default is "status"}
}
\value{
A list containing two tables
}
\description{
This function takes a data set of phs, age, and status and returns two
matrices: one is the PHS score at each of 99 percentiles (1-99\\%) and the other
is predicted K-M curve (plus confidence intervals) for each of these
percentiles. These two tables contain now individualized subject information
and can be exported to predict risk for new subjects for which you have PHS
scores.
}
\examples{

lookup_table <- create_lookup_table(test_data)

}
