% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_inc.R
\name{cum_inc}
\alias{cum_inc}
\title{Compute Cumulative Incidence}
\usage{
cum_inc(
  country,
  inc_file = "prostate_cancer_incidence_country.csv",
  prop_file = "intermediate_high_proportions_CAP.csv"
)
}
\arguments{
\item{country}{The country name}

\item{inc_file}{The incidence file name}

\item{prop_file}{The proportions file name}
}
\value{
A list with cumulative incidence and incidence data frames
}
\description{
Compute Cumulative Incidence
}
\examples{
cuminc_data <- cum_inc("USA")
}
