% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{Returns a concordance index}
\usage{
get_ci(
  data = NULL,
  phs = "phs",
  age = "age",
  status = "status",
  upper_quantile = 0.8,
  lower_quantile = 0.2,
  swc = FALSE,
  swc_popnumcases = NULL,
  swc_popnumcontrols = NULL,
  boot = FALSE,
  B = 1000
)
}
\arguments{
\item{data}{an optional data.frame containing the variables for phs, age, and status}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject or the unquoted name of a vector containing these values. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject or the unquoted name of a vector containing these values. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event) or the unquoted name of a vector containing these values. The default is "status"}

\item{swc}{logical. if \code{TRUE} performs sample weight correction}

\item{swc_popnumcases}{an optional integer specifying the number of cases in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{an optional integer specifying the number of controls in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{boot}{logical. if \code{TRUE} performs bootstrap and returns 95\% confidence intervals. Default = \code{FALSE}.}

\item{B}{Number of bootstrap iterations to run. Required if boot = \code{TRUE}. Default = 1000.}
}
\value{
A numeric hazard ratio or a list containing HR and the 95\% confidence intervals from bootstrap
}
\description{
This function returns the concordance index from a coxph fit and optionally
performs bootstrap resampling to return 95\% confidence intervals.
The data can either be provided as a data.frame with columns containing
the phs, age, and status of each individual or separate vectors containing
each of these values. The columns in \code{data} default to 'phs', 'age', and
'status', but any arbitrary column names can be used if named explicitly.
}
\examples{

CI <- get_ci(df, boot = TRUE, B = 300)

}
