% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_age.R
\name{get_phs_age}
\alias{get_phs_age}
\title{Polygenic Hazard Score Age Calculation}
\usage{
get_phs_age(age, ref, riskgroup, estimate_age)
}
\arguments{
\item{age}{Vector of ages}

\item{ref}{Reference scores}

\item{riskgroup}{Vector of risk group scores}

\item{estimate_age}{Age to estimate}
}
\value{
Estimated polygenic hazard score age
}
\description{
Polygenic Hazard Score Age Calculation
}
\examples{
\dontrun{
phs_age <- get_phs_age(age, ref, riskgroup, estimate_age)
}
}
