% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazrd-package.R
\docType{package}
\name{hazrd-package}
\alias{hazrd}
\alias{hazrd-package}
\title{hazrd: Polygenic Hazard Score with Cox Proportional Hazards Model}
\description{
Tools to evaluate and visualize Polygenic Hazard Score (PHS) analyses. Provides a consistent API for computing survival-based discrimination metrics (hazard ratio, C-index, odds ratio, HR per SD), bootstrapped confidence intervals, Kaplan-Meier curves, and cumulative incidence (absolute risk) curves stratified by PHS percentile group.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://amorris28.github.io/hazrd/}
  \item Report bugs at \url{https://github.com/amorris28/hazrd/issues}
}

}
\author{
\strong{Maintainer}: Andrew Morris \email{andrewmorris@mailbox.org}

}
\keyword{internal}
