% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phs_input_example}
\alias{phs_input_example}
\title{PHS Input Example}
\format{
\subsection{\code{phs_input_example}}{

A data frame with 4,000 rows and 1,996 columns:
\describe{
\item{age}{Age of diagnosis or age of censoring for each subject}
\item{status}{Case/control status}
\item{c11_pos1707163, ...}{Genotype at each locus}
}
}
}
\usage{
phs_input_example
}
\description{
A set of random genotype data
}
\keyword{datasets}
