% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsHist.R
\name{phsHist}
\alias{phsHist}
\title{Plot Histogram of Cases and Control by PHS}
\usage{
phsHist(model_file, metadata_file, inverse)
}
\arguments{
\item{model_file}{Path to the PHS file (tab delimited, no header, first column ID, second column PHS score)}

\item{metadata_file}{Path to the metadata file (tab delimited with column names bfile_id, age, status)}

\item{inverse}{Boolean indicating whether to inverse (x * -1) the PHS scores to reverse the direction of effect.}
}
\value{
A ggplot object
}
\description{
This function takes PHS scores and case/control status and plots a histogram..
}
\examples{
phs_hist <- phsHist(model_file, metadata_file, inverse)
}
