% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_phs.R
\name{reg_phs}
\alias{reg_phs}
\title{Regularized Polygenic Hazard Score}
\usage{
reg_phs(Age, status, Xvars)
}
\arguments{
\item{Age}{Vector of ages}

\item{status}{Vector indicating event status (0 = censored, 1 = event)}

\item{Xvars}{Matrix of predictors}
}
\value{
Vector of coefficients from the lasso-regularized Cox regression model
}
\description{
Regularized Polygenic Hazard Score
}
\examples{
beta <- reg_phs(Age, status, Xvars)
}
