% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerf.R
\name{RK_get_perf}
\alias{RK_get_perf}
\title{Performance Metrics for Polygenic Hazard Score}
\usage{
RK_get_perf(lp, Age, status, ref = NULL, swc_switch = TRUE, swc_caco = NULL)
}
\arguments{
\item{lp}{Linear predictor}

\item{Age}{Age vector}

\item{status}{Status vector (0 = censored, 1 = event)}

\item{ref}{Reference values (optional)}

\item{swc_switch}{Boolean to enable weighting}

\item{swc_caco}{Vector indicating case control status (optional)}
}
\value{
A list of performance metrics
}
\description{
Performance Metrics for Polygenic Hazard Score
}
\examples{
perf_metrics <- RK_get_perf(lp, Age, status)
}
