% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsAge.R
\name{RK_get_phsAge}
\alias{RK_get_phsAge}
\title{Polygenic Hazard Score Age Calculation}
\usage{
RK_get_phsAge(age, ref, riskgroup, estimate_age)
}
\arguments{
\item{age}{Vector of ages}

\item{ref}{Reference scores}

\item{riskgroup}{Vector of risk group scores}

\item{estimate_age}{Age to estimate}
}
\value{
Estimated polygenic hazard score age
}
\description{
Polygenic Hazard Score Age Calculation
}
\examples{
phs_age <- RK_get_phsAge(age, ref, riskgroup, estimate_age)
}
