% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPHS.R
\name{RK_get_regPHS}
\alias{RK_get_regPHS}
\title{Regularized Polygenic Hazard Score}
\usage{
RK_get_regPHS(Age, status, Xvars)
}
\arguments{
\item{Age}{Vector of ages}

\item{status}{Vector indicating event status (0 = censored, 1 = event)}

\item{Xvars}{Matrix of predictors}
}
\value{
Vector of coefficients from the lasso-regularized Cox regression model
}
\description{
Regularized Polygenic Hazard Score
}
\examples{
beta <- RK_get_regPHS(Age, status, Xvars)
}
