% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{PHS Test Data}
\format{
\subsection{\code{test_data}}{

A data frame with 1,000 rows and 3 columns:
\describe{
\item{phs}{Polygenic hazard score}
\item{age}{Age of diagnosis or age of censoring for each subject}
\item{status}{Case/control status}
}
}
}
\usage{
test_data
}
\description{
A simulated dataset of PHS scores
}
\keyword{datasets}
