% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_conf.R
\name{boot_conf}
\alias{boot_conf}
\title{Bootstrap confidence intervals for a custom statistic}
\usage{
boot_conf(df, bootstrap.iterations, conf.level = 0.95, f, ...)
}
\arguments{
\item{df}{a data.frame containing the columns phs, age, and status}

\item{bootstrap.iterations}{number of bootstrap iterations to run.}

\item{conf.level}{The confidence level to use for the confidence interval. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{f}{the name of a function to perform bootstrapping on.}

\item{...}{arguments to be passed on to \code{f()}.}
}
\value{
A list of confidence intervals
}
\description{
This function allows advanced users to compute bootstrap confidence
intervals for any statistic so long as the function passed to \code{f}
returns a single value of that statistic. For most users, the built-in
functions such as \code{\link{get_hr}} are recommended.
}
