% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cindex.R
\name{calc_cindex}
\alias{calc_cindex}
\title{Returns the concordance index for a dataset using a coxph fit}
\usage{
calc_cindex(df)
}
\arguments{
\item{df}{a data.frame containing the columns phs, age, and status}
}
\value{
A numeric hazard ratio
}
\description{
Internal function. Not intended for users.
}
