% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cindex.R
\name{calc_cindex}
\alias{calc_cindex}
\title{Returns the concordance index for a dataset using a coxph fit}
\usage{
calc_cindex(df, swc = FALSE, swc_popnumcases = NULL, swc_popnumcontrols = NULL)
}
\arguments{
\item{df}{a data.frame containing the clumnes phs, age, and status}

\item{swc}{logical. if \code{TRUE} performs sample weight correction}

\item{swc_popnumcases}{an optional integer specifying the number of cases in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{an optional integer specifying the number of controls in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{upper_quantile}{a vector specifying the upper quantile of the hazard ratio. Can also be supplied as a vector of length 2 to specify both the upper and lower limits of the quantile (e.g., \code{c(0.80, 0.98)}). If only one value is provided, then the PHS scores between that number and Infinite are included. The default is \code{0.80}.}

\item{lower_quantile}{a vector specifying the lower quantile of the hazard ratio. Can also be supplied as a vector of length 2 to specify both the upper and lower limits of the quantile (e.g., \code{c(0.3, 0.7)}). If only one value is provided, then the PHS scores between -Infinite and that number are included. The default is \code{0.20}.}
}
\value{
A numeric hazard ratio
}
\description{
Internal function. Not intended for users.
}
\examples{

c_index <- calc_cindex(df)

}
