% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cindex.R
\name{get_cindex}
\alias{get_cindex}
\alias{get_cindex.data.frame}
\alias{get_cindex.phsfit}
\title{Return concordance index (C-index) from a PHS model}
\usage{
get_cindex(object, ...)

\method{get_cindex}{data.frame}(
  object,
  phs = "phs",
  age = "age",
  status = "status",
  bootstrap.iterations = NULL,
  conf.level = 0.95,
  ...
)

\method{get_cindex}{phsfit}(object, bootstrap.iterations = NULL, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An object for which a concordance index is desired. Typically a
data.frame or a \code{phsfit} object.}

\item{...}{Additional arguments (currently ignored).}

\item{phs}{Optional string specifying the column in \code{data} containing polygenic hazard scores. Default: "phs".}

\item{age}{Optional string specifying the column in \code{data} containing ages. Default: "age".}

\item{status}{Optional string specifying the column in \code{data} containing case-control status (0=censored, 1=event). Default: "status".}

\item{bootstrap.iterations}{Number of bootstrap iterations (used only in data.frame method or if not precomputed in phsfit).}

\item{conf.level}{Confidence level for intervals (default 0.95).}
}
\value{
A list containing:
\describe{
\item{value}{Point estimate of the concordance index}
\item{conf.low}{Lower bound of confidence interval (if bootstrapped)}
\item{conf.high}{Upper bound of confidence interval (if bootstrapped)}
}
}
\description{
\code{get_cindex()} calculates the concordance index from either a fitted PHS object
(\code{phsfit}) or directly from a data.frame. Optionally, bootstrap confidence intervals
can be computed.
}
\examples{
# Using a data.frame
c_index <- get_cindex(test_data, bootstrap.iterations = 300)

# Using a phsfit object
phsfit <- fit_phs(test_data, bootstrap.iterations = 300)
c_index <- get_cindex(phsfit)

}
