% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hr.R
\name{get_hr}
\alias{get_hr}
\title{Return hazard ratio for Polygenic Hazard Score}
\usage{
get_hr(
  phs,
  age,
  status,
  data = NULL,
  upper_quantile = 0.8,
  lower_quantile = 0.2,
  swc = FALSE,
  swc_popnumcases = NULL,
  swc_popnumcontrols = NULL
)
}
\arguments{
\item{phs}{a vector of polygenic hazard scores for subjects or a string specifying the column name in \code{data} containing these values}

\item{age}{a vector of ages or a string specifying the column name in \code{data} containing these values. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation).}

\item{status}{a vector of case-control status (0 = censored, 1 = event) or a string specifying the column name in \code{data} containing these values}

\item{data}{an optional data.frame containing the variables phs, age, and status}

\item{upper_quantile}{an optional vector specifying the upper quantile of the hazard ratio. The default is \code{0.80}.}

\item{lower_quantile}{an optional vector specifying the lower quantile of the hazard ratio. The default is \code{0.20}.}

\item{swc}{logical. if \code{TRUE} performs sample weight correction}

\item{swc_popnumcases}{an optional integer specifying the number of cases in a reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{an optional integer specifying the number of controls in a reference population for sample weight correction. Required if swc = \code{TRUE}.}
}
\value{
A numeric hazard ratio
}
\description{
Return hazard ratio for Polygenic Hazard Score
}
\examples{
HR80_20 <- get_hr(phs, age, status)
}
