% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hr.R
\name{get_hr}
\alias{get_hr}
\title{Return hazard ratio for Polygenic Hazard Score}
\usage{
get_hr(
  phs,
  age,
  status,
  upper_quantile = 0.8,
  lower_quantile = 0.2,
  swc = FALSE,
  swc_popnumcases = NULL,
  swc_popnumcontrols = NULL
)
}
\arguments{
\item{phs}{Polygenic hazard score vector}

\item{age}{Age vector}

\item{status}{Status vector (0 = censored, 1 = event)}

\item{upper_quantile}{Status vector (0 = censored, 1 = event)}

\item{lower_quantile}{Status vector (0 = censored, 1 = event)}

\item{swc}{Boolean to enable sample weight correction}

\item{swc_popnumcases}{Number of cases in reference population for sample weight correction. Required if swc = \code{TRUE}.}

\item{swc_popnumcontrols}{Number of controls in reference population for sample weight correction. Required if swc = \code{TRUE}.}
}
\value{
A numeric hazard ratio
}
\description{
Return hazard ratio for Polygenic Hazard Score
}
\examples{
HR80_20 <- get_hr(phs, age, status)
}
