% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hr.R
\name{get_hr}
\alias{get_hr}
\alias{get_hr.data.frame}
\alias{get_hr.phsfit}
\title{Return the hazard ratio (HR) from a PHS model}
\usage{
get_hr(object, ...)

\method{get_hr}{data.frame}(
  object,
  phs = "phs",
  age = "age",
  status = "status",
  numerator = c(0.8, 1),
  denominator = c(0, 0.2),
  bootstrap.iterations = NULL,
  conf.level = 0.95,
  ...
)

\method{get_hr}{phsfit}(
  object,
  numerator = c(0.8, 1),
  denominator = c(0, 0.2),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{object}{a \code{phsfit} object}

\item{...}{Additional arguments (currently ignored)}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event). The default is "status"}

\item{numerator}{a vector specifying the quantiles of the numerator (e.g., \code{c(0.80, 0.98)}). The default is \code{c(0.8, 1.0)}.}

\item{denominator}{a vector specifying the quantiles of the denominator (e.g., \code{c(0.30, 0.70)}). The default is \code{c(0.0, 0.2)}.}

\item{bootstrap.iterations}{Number of bootstrap iterations to run to generate confidence intervals.}

\item{conf.level}{The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.}
}
\value{
A list.

A numeric hazard ratio or a list containing HR and the 95\% confidence intervals from bootstrap

A numeric hazard ratio or a list containing HR and the 95\% confidence intervals from bootstrap
}
\description{
\code{get_hr()} is a generic function that calculate the HR from a coxph fit or
directly from a data frame with specified columns. It returns a list with
the HR point estimate and optional bootstrap confidence intervals

This function calculates the hazard ratio and optionally performs
bootstrap resampling to return 95\% confidence intervals.
The data can either be provided as a data.frame with columns containing
the phs, age, and status of each individual or separate vectors containing
each of these values. The columns in \code{data} default to 'phs', 'age', and
'status', but any arbitrary column names can be used if named explicitly.

This function calculates the hazard ratio and optionally performs
bootstrap resampling to return 95\% confidence intervals.
The data can either be provided as a data.frame with columns containing
the phs, age, and status of each individual or separate vectors containing
each of these values.
}
\examples{
# phsfit interface:
\dontrun{
    get_hr(phsfit)
}
# Data frame interface:
get_hr(test_data, age = "age", status = "status", phs = "phs")

HR80_20 <- get_hr(test_data, bootstrap.iterations = 300)


HR80_20 <- get_hr(test_data, bootstrap.iterations = 300)

}
\seealso{
\code{\link[=get_hr.phsfit]{get_hr.phsfit()}}, \code{\link[=get_hr.data.frame]{get_hr.data.frame()}}
}
