% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or.R
\name{get_or}
\alias{get_or}
\title{Return odds ratio for Polygenic Hazard Score at a particular age}
\usage{
get_or(
  data = NULL,
  phs = "phs",
  age = "age",
  status = "status",
  or_age,
  numerator = c(0.8, 1),
  denominator = c(0, 0.2),
  bootstrap.iterations = NULL,
  conf.level = 0.95
)
}
\arguments{
\item{data}{an optional data.frame containing the variables for phs, age, and status}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event). The default is "status"}

\item{or_age}{an integer specifying the age at which the odds ratio should be calculated}

\item{numerator}{a vector specifying the quantiles of the numerator (e.g., \code{c(0.80, 0.98)}). The default is \code{c(0.8, 1.0)}.}

\item{denominator}{a vector specifying the quantiles of the denominator (e.g., \code{c(0.30, 0.70)}). The default is \code{c(0.0, 0.2)}.}

\item{bootstrap.iterations}{Number of bootstrap iterations to run.}

\item{conf.level}{The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.}
}
\value{
A numeric odds ratio
}
\description{
This function calculate the odds ratio at a particular age.
The data can either be provided as a data.frame with columns containing
the phs, age, and status of each individual or separate vectors containing
each of these values. The columns in \code{data} default to 'phs', 'age', and
'status', but any arbitrary column names can be used if named explicitly.
}
\examples{

OR80_20 <- get_or(test_data, or_age = 70)

}
