% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or.R
\name{get_or}
\alias{get_or}
\title{Return odds ratio for Polygenic Hazard Score at a particular age}
\usage{
get_or(
  data = NULL,
  phs = "phs",
  age = "age",
  status = "status",
  or_age,
  lower_interval = 0.2,
  upper_interval = 0.8,
  CI = FALSE,
  bootstrap_iterations = 1000
)
}
\arguments{
\item{data}{an optional data.frame containing the variables for phs, age, and status}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject or the unquoted name of a vector containing these values. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject or the unquoted name of a vector containing these values. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event) or the unquoted name of a vector containing these values. The default is "status"}

\item{or_age}{an integer specifying the age at which the odds ratio should be calculated}

\item{lower_interval}{a vector specifying the quantiles of the lower interval. If a single value is given, that will be used as the upper quantile and the lower quantile will be \code{-Inf}. If a vector of length 2 is provided then these will be used as the lower and upper quantiles of the interval (e.g., \code{c(0.30, 0.70)}). The default is \code{0.2}.}

\item{upper_interval}{a vector specifying the quantiles of the upper interval. If a single value is given, that will be used as the lower quantile and the upper quantile will be \code{Inf}. If a vector of length 2 is provided then these will be used as the lower and upper quantiles of the interval (e.g., \code{c(0.80, 0.98)}). The default is \code{0.80}.}

\item{CI}{logical. If \code{TRUE} performs bootstrap and returns 95\% confidence intervals. Default = \code{FALSE}.}

\item{bootstrap_iterations}{Number of bootstrap iterations to run. Required if boot = \code{TRUE}. Default = 1000.}
}
\value{
A numeric odds ratio
}
\description{
This function calculate the odds ratio at a particular age.
The data can either be provided as a data.frame with columns containing
the phs, age, and status of each individual or separate vectors containing
each of these values. The columns in \code{data} default to 'phs', 'age', and
'status', but any arbitrary column names can be used if named explicitly.
}
\examples{
\dontrun{
OR80_20 <- get_or(data, or_age = 70)
}
}
