% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or.R
\name{get_or}
\alias{get_or}
\alias{get_or.data.frame}
\alias{get_or.phsfit}
\title{Return odds ratios (OR) from a PHS model at a specified age}
\usage{
get_or(object, or_age, ...)

\method{get_or}{data.frame}(
  object,
  or_age,
  phs = "phs",
  age = "age",
  status = "status",
  numerator = c(0.8, 1),
  denominator = c(0, 0.2),
  bootstrap.iterations = NULL,
  conf.level = 0.95,
  ...
)

\method{get_or}{phsfit}(
  object,
  or_age,
  numerator = c(0.8, 1),
  denominator = c(0, 0.2),
  bootstrap.iterations = NULL,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{object}{An object for which an odds ratio is desired. Typically a
data.frame or a \code{phsfit} object.}

\item{or_age}{Integer specifying the age at which the odds ratio should be calculated.}

\item{...}{Additional arguments (currently ignored)}

\item{phs}{an optional string specifying the column name in \code{data} containing the polygenic hazard score for each subject. The default is "phs"}

\item{age}{an optional string specifying the column name in \code{data} containing the age of each subject. For cases, this should be the age at event (e.g., diagnosis) and for controls this should be age of censoring (e.g., last observation). The default is "age"}

\item{status}{an optional string specifying the column name in \code{data} containing case-control status (0 = censored, 1 = event). The default is "status"}

\item{numerator}{Numeric vector of quantiles for the numerator (e.g., \code{c(0.80, 1.0)}).}

\item{denominator}{Numeric vector of quantiles for the denominator (e.g., \code{c(0.0, 0.2)}).}

\item{bootstrap.iterations}{Number of bootstrap iterations (used only in data.frame method).}

\item{conf.level}{Confidence level for intervals (default 0.95).}
}
\value{
A list containing:
\describe{
\item{value}{Point estimate of the OR}
\item{conf.low}{Lower bound of confidence interval (if bootstrapped)}
\item{conf.high}{Upper bound of confidence interval (if bootstrapped)}
}
}
\description{
\code{get_or()} calculates odds ratios from either a fitted PHS object (\code{phsfit})
or directly from a data.frame at a specified age. It returns a list containing
the point estimate and optionally bootstrap confidence intervals.
}
\examples{
# Using a data.frame
OR <- get_or(test_data, or_age = 70, bootstrap.iterations = 300)

# Using a phsfit object
phsfit <- fit_phs(test_data, bootstrap.iterations = 300)
OR <- get_or(phsfit, or_age = 70)

}
