% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazrd-package.R
\docType{package}
\name{hazrd-package}
\alias{hazrd}
\alias{hazrd-package}
\title{hazrd: Polygenic Hazard Score with Cox Proportional Hazards Model}
\description{
Tools to compute polygenic hazard scores using Cox proportional hazards model and lasso regularization via glmnet.
}
\author{
\strong{Maintainer}: Andrew Morris \email{andrewmorris@mailbox.org}

}
\keyword{internal}
