% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_curve.R
\name{km_curve}
\alias{km_curve}
\title{Plot Kaplan-Meier curve}
\usage{
km_curve(model_file, metadata_file, inverse = FALSE, ideal = FALSE)
}
\arguments{
\item{model_file}{Path to the PHS file (tab delimited, no header, first column ID, second column PHS score)}

\item{metadata_file}{Path to the metadata file (tab delimited with column names bfile_id, age, status)}

\item{inverse}{Boolean indicating whether to inverse (x * -1) the PHS scores to reverse the direction of effect.}

\item{ideal}{Boolean indicating whether to add the idealized survival curve.}
}
\value{
A ggplot2 plot of a K-M curve
}
\description{
This function processes PHS data and generates a K-M plot.
}
\examples{
km_curve <- km_curve(model_file, metadata_file, inverse = TRUE, ideal = FALSE)
}
