% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_hist.R
\name{phs_hist}
\alias{phs_hist}
\title{Plot Histogram of Cases and Control by PHS}
\usage{
phs_hist(df, normalize = FALSE, scale = FALSE, inverse = FALSE)
}
\arguments{
\item{df}{Data.frame or tibble containing a column for \code{phs} and a column for \code{status}. All other columns ignored.}

\item{normalize}{Boolean indicating whether to normalize the histogram to area = 1. This is replaces the y axis with denisty instead of count. Helpful to compare distributions with very different counts. Default = \code{FALSE}.}

\item{scale}{Boolean indicating whether to center and scale the PHS scores to unit variance. Default = \code{FALSE}.}

\item{inverse}{Boolean indicating whether to inverse (x * -1) the PHS scores to reverse the direction of effect. Default = \code{FALSE}.}
}
\value{
A ggplot object
}
\description{
This function takes PHS scores and case/control status and plots a histogram.
}
\examples{
phs_hist <- phs_hist(df)
}
